+++ sdk/formrecognizer/ai-form-recognizer/src/constants.ts
@@ -2,10 +2,22 @@
 // Licensed under the MIT license.
 
 /**
+ * Defines known cloud audiences for Form Recognizer.
+ */
+export enum KnownFormRecognizerAudience {
+  /** Azure China */
+  AzureResourceManagerChina = "https://cognitiveservices.azure.cn/.default",
+  /** Azure Government */
+  AzureResourceManagerGovernment = "https://cognitiveservices.azure.us/.default",
+  /** Azure Public Cloud */
+  AzureResourceManagerPublicCloud = "https://cognitiveservices.azure.com/.default",
+}
+
+/**
  * The default AAD permissions scope for Cognitive Services.
  * @internal
  */
-export const DEFAULT_COGNITIVE_SCOPE = "https://cognitiveservices.azure.com/.default";
+export const DEFAULT_COGNITIVE_SCOPE = KnownFormRecognizerAudience.AzureResourceManagerPublicCloud;
 
 /**
  * @internal